Add an authorization policy
And a forbidden view for the API.

This means that API views can use `permission="foo"` in the
`@view_config` and if the request doesn't have the necessary permission
the view won't be called and Pyramid will 403 the request instead.

It's no longer necessary to add an "unauthorized" version of each API
view to set the error response's status code.

A Pyramid forbidden view is added to change API 403 responses into JSON.
