+++ tests/functional/checkmate/views/api/check_url_test.py
@@ -1,15 +1,15 @@
 class TestAPIAuth:
-    def test_401_unauthenticated(self, app):
+    def test_403_forbidden(self, app):
         res = app.get("/api/check", expect_errors=True)
 
-        assert res.status_code == 401
+        assert res.status_code == 403
 
-    def test_401_unauthenticated_no_matching_users(self, app):
+    def test_403_forbidden_no_matching_users(self, app):
         app.authorization = ("Basic", ("api_nonsense",))
 
         res = app.get("/api/check", expect_errors=True)
 
-        assert res.status_code == 401
+        assert res.status_code == 403
 
     def test_authenticated_existing_user(self, app):
         app.authorization = ("Basic", ("dev_api_key", ""))
