+++ tests/unit/checkmate/views/api/exception_test.py
@@ -1,9 +1,10 @@
 from unittest.mock import create_autospec
 
 import pytest
+from pyramid.httpexceptions import HTTPForbidden
 
 from checkmate.exceptions import JSONAPIException
-from checkmate.views.api.exception import api_error
+from checkmate.views.api.exception import api_error, api_forbidden
 
 
 class TestAPIError:
@@ -18,3 +19,11 @@ class TestAPIError:
         error = create_autospec(JSONAPIException, spec_set=True, instance=True)
         error.status_code = 401
         return error
+
+
+class TestAPIForbidden:
+    def test_it(self, pyramid_request):
+        response = api_forbidden(HTTPForbidden(), pyramid_request)
+
+        assert pyramid_request.response.status_int == HTTPForbidden.code
+        assert response == {"error": "invalid_token"}
